+++ src/main/java/testmain/SimplePP.java
@@ -16,10 +16,10 @@ public class SimplePP {
         this.pathLength = pathLength;
     }
 
-    public ItemPosition getPointAtDistance(ItemPosition i, int xdist,int ydist, int zdist) {
-         int x = 0;
-         int y = 0;
-         int z = 0;
+    public ItemPosition getPointAtDistance(ItemPosition i, double xdist,double ydist, double zdist) {
+         double x = 0;
+         double y = 0;
+         double z = 0;
          Random ran = new Random();
 
          if (this.destination.x > i.x) {
@@ -57,16 +57,16 @@ public class SimplePP {
         Stack<ItemPosition> temp = new Stack<ItemPosition>();
 
         temp.push(this.start);
-        int xdist = Math.abs(this.start.x - this.destination.x);
-        int ydist = Math.abs(this.start.y - this.destination.y);
-        int zdist = Math.abs(this.start.z - this.destination.z);
+        double xdist = Math.abs(this.start.x - this.destination.x);
+        double ydist = Math.abs(this.start.y - this.destination.y);
+        double zdist = Math.abs(this.start.z - this.destination.z);
         //System.out.println("xdist: " + xdist + " ydist: " + ydist + " zdis: " + zdist);
 
         //int xincrement = xdist/pathLength;
         //int yincrement = ydist/pathLength;
-        int xincrement = 0;
-        int yincrement = 0;
-        int zincrement = zdist/pathLength;
+        double xincrement = 0;
+        double yincrement = 0;
+        double zincrement = zdist/pathLength;
         //System.out.println("xinc: " + xincrement + " yinc: " + yincrement + " zinc: " + zincrement);
         
         ItemPosition i = this.start;
